moves endpoints from providerPrograms to providers
